requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-openai
pydantic>=2.0
tavily-python
firecrawl-py
python-dotenv
chromadb
sentence-transformers
python-multipart
